Implement configuration changes
The real goal here was to implement the usage of a configuration setting to determine which database should be used for access and to create the correct connection.

Second was to enable unit tests to successfully complete for both LISP and LISPROD.

Lastly, the SQL for the RP_ACCTS Get() method was simplified.