Added IComparer(), added missing LIS schema reference, simplified DB access, added helper methods to ComparisonTests.
Added implementation of IComparer to BaseTableClass to allow comparison of property values of one object instance to another;

Added LIS schema reference to MSSQL queries.

Simplified Get() methods for ACCT_TAGS and RP_ACCTS to use the IDbConnection passed in.

Added helper methods to ComparisonTests to compare objects and lists of objects. If object comparison fails, the ToString() method is called to dump the properties.